# gateway/internal/swagger/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Habitizer Microservices API Ecosystem",
    "description": "Comprehensive REST API specification for Habitizer: Behavioral Psychology & Habit Substitution Platform.\n\n### Microservices Architecture:\n- **Gateway Service** (Port 8000): Unified reverse proxy, clean static hosting, and live SSE test streaming.\n- **Auth Service** (Port 8001): User registration, JWT authentication, and profile token verification.\n- **Habit Service** (Port 8002): Habit substitution loop CRUD, nightly check-in evaluation, suggestions catalog, auto-promotion, and Google Calendar auto-scheduling.\n- **Analytics Service** (Port 8003): Streak calculation, freeze ledger, coin economy, and reward redemptions.",
    "version": "1.0.0",
    "contact": {
      "name": "Habitizer Engineering Team",
      "email": "support@habitizer.io"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "API Gateway (Local Development)"
    }
  ],
...
  "paths": {
...
    "/api/v1/habits": {
      "get": {
        "tags": ["Habits & Substitution Loops"],
        "summary": "List all habits for user",
        "description": "Retrieves the active habit substitution loops for the authenticated user.",
        "parameters": [
          {
            "name": "user_id",
            "in": "query",
            "required": false,
            "schema": { "type": "string" },
            "description": "Optional User ID filter"
          }
        ],
        "responses": {
          "200": {
            "description": "List of habit entities",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Habit"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Habits & Substitution Loops"],
        "summary": "Create a new Habit Substitution Loop",
        "description": "Creates a new habit loop containing Cue Trigger, Bad Habit, Healthy Routine, and Neuro Reward.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateHabitRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Habit created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Habit"
                }
              }
            }
          },
          "400": {
            "description": "Free tier limit reached (max 3 habits) or invalid payload",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Habit": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "example": "hab_47c2e891-b11d-4074-bcf2-998877665544" },
          "user_id": { "type": "string", "example": "usr_123" },
          "bad_habit": { "type": "string", "example": "Late Night Junk Food Snacking" },
          "cue_trigger": { "type": "string", "example": "Stress or boredom around 11:00 PM" },
          "replacement_habit": { "type": "string", "example": "5-Minute Deep Breathing & Herbal Tea" },
          "neuro_reward": { "type": "string", "example": "15 Mins Clean Screen Time Pass" },
          "scheduled_time": { "type": "string", "example": "21:30" },
          "frequency": { "type": "string", "example": "daily" },
          "is_active": { "type": "boolean", "example": true },
          "created_at": { "type": "string", "format": "date-time" }
        }
      },
      "CreateHabitRequest": {
        "type": "object",
        "required": ["bad_habit", "cue_trigger", "replacement_habit"],
        "properties": {
          "user_id": { "type": "string", "example": "usr_123" },
          "bad_habit": { "type": "string", "example": "Mindless Social Scrolling" },
          "cue_trigger": { "type": "string", "example": "Waking up in the morning" },
          "replacement_habit": { "type": "string", "example": "10-Minute Morning Stretches & Water" },
          "neuro_reward": { "type": "string", "example": "Speciality Espresso Coffee" },
          "scheduled_time": { "type": "string", "example": "08:00" },
          "frequency": { "type": "string", "example": "daily" }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": false },
          "error": { "type": "string", "example": "Invalid request parameter" }
        }
      }
    }
  }
}